+++ tests/test_vector.py
@@ -140,6 +140,7 @@ def test_sets():
     assert vec.maximum(other).equal(vector({0: 1.0, 1: 1.0, 2: 2.0}))
     assert vec.minimum(other).equal(vector({0: 0.0, 1: 0.0, 2: 1.0}))
     assert (vec ^ other).equal(vector({0: 1.0, 3: 1.0}))
+    assert (vector() ^ vec).equal(vec)
     assert vec.difference(other, ()).equal(vector({0: 1.0}))
 
     vec |= other