+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * Python >=3.11 required
 * Stable ABI builds
-* Copy and `xor` optimized
+* Copy, conversion, and `xor` optimized
 
 ### Added
 * `toarrays` method