+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+
+## [1.7](https://pypi.org/project/spector/1.7/) - 2026-07-20
 ### Changed
 * Python >=3.11 required
 * Stable ABI builds